requirements.txt
requests==2.31.0
aiohttp==3.8.5
numpy==1.24.3
pandas==2.0.3
torch==2.0.1
transformers==4.34.0
accelerate==0.23.0
huggingface-hub==0.17.3
datasets==2.14.5
# Additional packages for text processing
arabic-reshaper==3.0.0
python-bidi==0.4.2
nltk==3.8.1
# Document processing
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
